requirements.txt
customtkinter
requests
Pillow
python-dotenv
